Add QuiltLoaderInternal to all of the plugin API.
