+++ src/main/java/org/quiltmc/loader/api/plugin/solver/AliasedLoadOption.java
@@ -17,8 +17,11 @@
 package org.quiltmc.loader.api.plugin.solver;
 
 import org.jetbrains.annotations.Nullable;
+import org.quiltmc.loader.impl.util.QuiltLoaderInternal;
+import org.quiltmc.loader.impl.util.QuiltLoaderInternalType;
 
 /** Indicates that this {@link LoadOption} should use {@link #getTarget()} for mod solving instead of this. */
+@QuiltLoaderInternal(QuiltLoaderInternalType.PLUGIN_API)
 public interface AliasedLoadOption {
 
     /** @return The {@link LoadOption} to use instead of this, or null if this can be used (and so this is not actually
