+++ src/main/java/org/quiltmc/loader/api/plugin/solver/TentativeLoadOption.java
@@ -16,10 +16,14 @@
 
 package org.quiltmc.loader.api.plugin.solver;
 
+import org.quiltmc.loader.impl.util.QuiltLoaderInternal;
+import org.quiltmc.loader.impl.util.QuiltLoaderInternalType;
+
 /** {@link LoadOption}s can implement this if they must be processed before they can be used, if they are selected.
  * <p>
  * Unselected {@link TentativeLoadOption}s are left alone at the end of the cycle, and are not resolved.
  */
+@QuiltLoaderInternal(QuiltLoaderInternalType.PLUGIN_API)
 public interface TentativeLoadOption {
 
 }
