+++ src/main/java/org/quiltmc/loader/impl/metadata/qmj/ProvidedModImpl.java
@@ -16,8 +16,8 @@
 
 package org.quiltmc.loader.impl.metadata.qmj;
 
+import org.quiltmc.loader.api.ModMetadata.ProvidedMod;
 import org.quiltmc.loader.api.Version;
-import org.quiltmc.loader.api.plugin.ModMetadataExt.ProvidedMod;
 import org.quiltmc.loader.impl.util.QuiltLoaderInternal;
 import org.quiltmc.loader.impl.util.QuiltLoaderInternalType;
