+++ src/main/java/org/quiltmc/loader/impl/metadata/qmj/QuiltModMetadataWrapperFabric.java
@@ -16,18 +16,26 @@
 
 package org.quiltmc.loader.impl.metadata.qmj;
 
-import java.util.*;
+import java.util.ArrayList;
+import java.util.Collection;
+import java.util.Collections;
+import java.util.HashMap;
+import java.util.List;
+import java.util.Map;
+import java.util.Optional;
 
 import org.jetbrains.annotations.Nullable;
 import org.quiltmc.loader.api.LoaderValue;
 import org.quiltmc.loader.api.ModContainer;
 import org.quiltmc.loader.api.ModContributor;
 import org.quiltmc.loader.api.ModLicense;
-import org.quiltmc.loader.api.plugin.ModMetadataExt.ProvidedMod;
+import org.quiltmc.loader.api.ModMetadata.ProvidedMod;
 import org.quiltmc.loader.impl.fabric.metadata.CustomValueImpl;
 import org.quiltmc.loader.impl.fabric.metadata.MapBackedContactInformation;
 import org.quiltmc.loader.impl.fabric.metadata.SimplePerson;
-import org.quiltmc.loader.impl.metadata.*;
+import org.quiltmc.loader.impl.metadata.EntrypointMetadata;
+import org.quiltmc.loader.impl.metadata.FabricLoaderModMetadata;
+import org.quiltmc.loader.impl.metadata.NestedJarEntry;
 import org.quiltmc.loader.impl.util.QuiltLoaderInternal;
 import org.quiltmc.loader.impl.util.QuiltLoaderInternalType;
